services/patec-backend/pyproject.toml
[project]
name = "patec-backend"
version = "0.1.0"
description = "PATEC - Parecer Tecnico de Engenharia - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "greenlet>=3.0.0",
    "email-validator>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "reportlab>=4.1.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: integration tests requiring database",
]
